+++ components/SearchForm/HelpSection.tsx
@@ -1,0 +1,23 @@
+export const HelpSection = () => (
+	<div className="mt-3 p-4 bg-blue-50 border border-blue-200 rounded-lg">
+		<h4 className="font-semibold text-blue-900 mb-2">
+			So findest du den Text zum Teilen:
+		</h4>
+		<ol className="text-sm text-blue-800 space-y-1 list-decimal list-inside">
+			<li>
+				Gehe auf <strong>bahn.de</strong> und plane deine Verbindung
+			</li>
+			<li>Wähle deine gewünschte Verbindung aus</li>
+			<li>
+				Klicke auf die drei kleinen Punkte oben in der Ecke der Verbindung
+			</li>
+			<li>
+				Klicke auf <strong>"Verbindung Teilen"</strong>
+			</li>
+			<li>
+				Klicke auf <strong>"Infos Kopieren"</strong>
+			</li>
+			<li>Füge den kompletten kopierten Text hier ein</li>
+		</ol>
+	</div>
+);
